requirements.txt
beautifulsoup4==4.11.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cohere==2.9.1
colorama==0.4.6
confection==0.0.3
cssselect==1.2.0
cymem==2.0.7
distlib==0.3.6
docopt==0.6.2
en-core-web-sm==3.4.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
flashtext==2.7
Flask==2.1.2
future==0.16.0
idna==3.4
importlib-metadata==5.1.0
inscriptis==2.3.1
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
lxml==4.9.1
MarkupSafe==2.1.1
murmurhash==1.0.9
networkx==2.6.3
newspaper3k==0.2.8
nltk==3.7
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pathy==0.10.0
Pillow==9.3.0
pipreqs==0.4.11
pke==2.0.0
platformdirs==2.5.4
preshed==3.0.8
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2018.9
PyYAML==6.0
Questgen==1.0.0
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.4.1
sense2vec==2.0.0
sentencepiece==0.1.97
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0.post1
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
strsim==0.0.3
thinc==8.1.5
threadpoolctl==3.1.0
tinysegmenter==0.3
tldextract==3.4.0
tokenizers==0.8.1rc1
torch==1.10.0
tqdm==4.64.1
transformers==3.0.2
typer==0.7.0
typing-extensions==4.1.1
Unidecode==1.3.4
urllib3==1.26.13
virtualenv==20.16.7
wasabi==0.10.1
Werkzeug==2.2.2
yarg==0.1.9
zipp==3.10.0
